replaced some tests
